Fix auth hydration redirects
