+++ src/components/blocks/login-route-shell.tsx
@@ -1,0 +1,11 @@
+import { LoginForm } from "@/components/login-form"
+
+export function LoginRouteShell() {
+  return (
+    <div className="flex min-h-svh flex-col items-center justify-center bg-zinc-100 p-6 md:p-10">
+      <div className="w-full max-w-sm md:max-w-4xl">
+        <LoginForm />
+      </div>
+    </div>
+  )
+}
